removing unused _check_generated_test_files and moving TestPlanParser to fileIO
